Update heading size styling to work with v0.13.30
Some of the classes appear to have changed in the latest version, breaking heading font-sizes on most themes. I added this as additions since I'm sure entirely sure what's for the new editor, old editor, what's going on with mobile, etc, and didn't want to break compatibility.

This worked for me. Happy to alter it if there's a method to the selectors that I'm not seeing.

Cheers